revert decreasing writing/reading max int in bit array

+++ branches/sphinxbase-trie/src/libsphinxbase/lm/lm_trie_find.h
@@ -9,9 +9,9 @@
 unigram_t* unigram_find(unigram_t *u, word_idx word, node_range_t *next);
 
 uint8 lm_trie_find(
-    void *base, uint8 total_bits, uint8 key_bits, uint32 key_mask,
-    uint32 before_it, uint32 before_v,
-    uint32 after_it, uint32 after_v,
-    uint32 key, uint32 *out);
+    void *base, uint8 total_bits, uint8 key_bits, uint64 key_mask,
+    uint64 before_it, uint64 before_v,
+    uint64 after_it, uint64 after_v,
+    uint64 key, uint64 *out);
 
 #endif /* __LM_TRIE_FIND_H__ */
